Move Paging and RemoteMediation into the bookmarks module.
