+++ data/database/src/commonMain/kotlin/dev.avatsav.linkding/data.db/inject/DatabaseComponent.kt
@@ -1,8 +1,9 @@
-package dev.avatsav.linkding.data.db.sqldelight
+package dev.avatsav.linkding.data.db.inject
 
 import dev.avatsav.linkding.data.db.Database
 import dev.avatsav.linkding.data.db.daos.BookmarksDao
 import dev.avatsav.linkding.data.db.daos.PagingBookmarksDao
+import dev.avatsav.linkding.data.db.sqldelight.DatabaseFactory
 import dev.avatsav.linkding.data.db.sqldelight.daos.SqlDelightBookmarksDao
 import dev.avatsav.linkding.data.db.sqldelight.daos.SqlDelightPagingBookmarksDao
 import dev.avatsav.linkding.inject.AppScope
@@ -11,12 +12,14 @@ import me.tatarka.inject.annotations.Provides
 expect interface SqlDelightPlatformComponent
 
 interface SqlDelightComponent : SqlDelightPlatformComponent {
-
     @AppScope
     @Provides
     fun provideSqlDelightDatabase(
         factory: DatabaseFactory,
     ): Database = factory.createDatabase()
+}
+
+interface DatabaseComponent : SqlDelightComponent {
 
     @AppScope
     @Provides
