+++ data/database/src/jvmMain/kotlin/dev/avatsav/linkding/data/db/inject/SqlDelightPlatformComponent.kt
@@ -1,7 +1,7 @@
-package dev.avatsav.linkding.data.db.sqldelight
+package dev.avatsav.linkding.data.db.inject
 
 import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.native.NativeSqliteDriver
+import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
 import dev.avatsav.linkding.data.db.Database
 import dev.avatsav.linkding.inject.AppScope
 import me.tatarka.inject.annotations.Provides
@@ -11,9 +11,8 @@ actual interface SqlDelightPlatformComponent {
     @Provides
     @AppScope
     fun provideSqliteDriver(): SqlDriver {
-        return NativeSqliteDriver(
-            schema = Database.Schema,
-            name = "linkding.db",
-        )
+        val driver: SqlDriver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
+        Database.Schema.create(driver)
+        return driver
     }
 }
